requirements.txt
# Core application dependencies
streamlit>=1.50.0
scikit-learn>=1.3.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
nltk>=3.8.0

# Production requirements
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == "win32"
python-multipart>=0.0.6
python-dotenv>=1.0.0
gunicorn>=21.2.0
uvicorn>=0.24.0

# Security and validation
cryptography>=41.0.0
pydantic>=2.0.0

# Monitoring and logging
psutil>=5.9.0

# Optional performance improvements
redis>=5.0.0